addressed review changes
