+++ app/src/androidTest/java/com/example/app/cases/MainActivityE2EWorkoutTests.kt
@@ -1,17 +1,62 @@
 package com.example.app.cases
 
+import android.content.Context
+import androidx.compose.ui.test.assertIsDisplayed
+import androidx.compose.ui.test.isDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.onNodeWithTag
+import androidx.compose.ui.test.onNodeWithText
+import androidx.compose.ui.test.performClick
+import androidx.compose.ui.test.performTextInput
 import bg.zahov.app.MainActivity
-import com.example.app.helpers.MainActivityRobot
+import bg.zahov.fitness.app.R
 import org.junit.Rule
 import org.junit.Test
+import kotlin.math.log
 
 class MainActivityE2EWorkoutTests {
 
     @get:Rule
     val composeTestRule = createAndroidComposeRule<MainActivity>()
+    private lateinit var context: Context
 
-    private lateinit var robot: MainActivityRobot
+    fun goToLoginScreen() {
+        composeTestRule.onNodeWithTag("to_login_screen")
+            .performClick()
+    }
+
+    fun login() {
+        composeTestRule.onNodeWithTag("to_home_screen").performClick()
+    }
+
+    fun inputCredentials(email: String = "test@gmail.com", password: String = "123456") {
+        composeTestRule.onNodeWithText(context.getString(R.string.email_text_field_hint))
+            .performTextInput(email)
+        composeTestRule.onNodeWithText(context.getString(R.string.password_text_field_hint))
+            .performTextInput(password)
+    }
+
+    fun checkIfInHomePage() {
+        composeTestRule.waitUntil(5000) {
+            composeTestRule.onNodeWithText(context.getString(R.string.dashboard)).isDisplayed()
+        }
+        composeTestRule.onNodeWithText(context.getString(R.string.dashboard)).assertIsDisplayed()
+    }
+
+    fun navigateToStartWorkout() {
+        composeTestRule.onNodeWithText(context.getString(R.string.workout)).performClick()
+    }
+
+    fun startWorkout() {
+        composeTestRule.onNodeWithText(context.getString(R.string.start_empty_workout_text))
+            .performClick()
+    }
+
+    fun verifyWorkout() {
+        composeTestRule.onNodeWithText(context.getString(R.string.add_note)).assertIsDisplayed()
+        composeTestRule.onNodeWithText(context.getString(R.string.add_exercise)).assertIsDisplayed()
+        composeTestRule.onNodeWithText(context.getString(R.string.cancel)).assertIsDisplayed()
+    }
 
     /**
      * Test the following scenario:
@@ -21,13 +66,14 @@ class MainActivityE2EWorkoutTests {
      */
     @Test
     fun testStartWorkout() {
-        robot = MainActivityRobot(composeTestRule, composeTestRule.activity.applicationContext)
-        robot.clickOnLoginButton()
-        robot.inputCredentials()
-        robot.clickOnLoginButton()
-        robot.checkIfInHomePage()
-        robot.navigateToStartWorkout()
-        robot.startWorkout()
-        robot.verifyWorkout()
+        context = composeTestRule.activity.applicationContext
+
+        goToLoginScreen()
+        inputCredentials()
+        login()
+        checkIfInHomePage()
+        navigateToStartWorkout()
+        startWorkout()
+        verifyWorkout()
     }
 }
